Support live removal of scheduled jobs
This is done by maintaining sets of currently active jobs and jobs that
are active but are scheduled for deletion. This allows adding and
removing jobs to be O(log n), at the cost of taking job execution from
O(1) to O(log n) since we need to check each job to see if it should be
deleted.